SearchHeader에 연결
