+++ src/App.jsx
@@ -1,16 +1,21 @@
 import { Route, Routes } from "react-router-dom";
-import Admin from "./pages/Admin/Admin";
-import Home from "./pages/Home/Home";
-import SignupPage from "../src/pages/SignupPage/SignupPage"
-import SigninPage from "../src/pages/SigninPage/SigninPage"
 import AdminRoute from "./Routes/AdminRoute";
 import MainRoute from "./Routes/MainRoute";
+import SinginPage from "./pages/SigninPage/SigninPage";
+import SignupPage from "./pages/SignupPage/SignupPage";
+import AdminSigninPage from "./pages/Admin/AdminSigninPage/AdminSigninPage";
 
 function App() {
   return (
     <>
-      <AdminRoute />
-      <MainRoute />
+      <Routes>
+        <Route path="/*" element={<MainRoute />}/>
+        <Route path="/userlogin" element={<SinginPage/> } />
+        <Route path="/usersignup" element={<SignupPage />} />
+
+        <Route path="/admin/*" element={<AdminRoute />}/>
+        <Route path="/adminlogin" element={<AdminSigninPage />}/>
+      </Routes>
     </>
   )
 }
