+++ src/Routes/AdminRoute.jsx
@@ -1,11 +1,16 @@
 import React from 'react';
-import { Routes } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
+import AdminMainPage from '../pages/Admin/AdminMainPage/AdminMainPage';
+import SinginPage from '../pages/SigninPage/SigninPage';
 
 function AdminRoute(props) {
     return (
-        <Routes>
-            
-        </Routes>
+        <>
+            <h1>어드민</h1>
+            <Routes>
+
+            </Routes>
+        </>
     );
 }
