--- src/components/Header/style.js
@@ -1,20 +1,0 @@
-import { css } from "@emotion/react";
-
-// 헤더
-export const headerLayout = css`
-    display: flex;
-    justify-content: center;
-    align-items: center;    
-    background-color: white;
-`;
-
-export const header = css`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 200px;
-    width: 1200px;
-    background-color: #e6dea8;
-    flex-direction: column;
-`;
-
