--- src/components/HeaderTop/style.js
@@ -1,15 +1,0 @@
-import { css } from "@emotion/react";
-
-export const headerTop = css`
-    box-sizing: border-box;
-    width: 100%;
-    height: 35px;
-    background-color: #33AFE9;
-`;
-
-export const button = css`
-    height: 100%;
-    border: none;
-    background-color: transparent;
-    cursor: pointer;
-`;
